+++ CHANGELOG.md
@@ -13,7 +13,7 @@ With this major release, the entire API has been rethought to allow the picker t
 - Added Grunt.js build system.
 - Added QUnit test suite.
 - Added Travis integration.
-- Updated themes to be Sass-based.
+- Updated themes to be LESS-based.
 - Removed “inline” and “inline-fixed” themes.
 - Removed jam.js bindings within `package.json`.
 - Removed official support for IE7. Still works but looks odd.